Split lambda implementation into several mods
